_4th_01_project/requirements.txt.txt
Django==5.2.5
langchain==0.3.27
langchain_core==0.3.74
langchain_openai==0.3.28
numpy==2.3.2
openai==1.99.3
paddleocr==3.1.0
paddlepaddle==3.1.0
Pillow==11.3.0
python-dotenv==1.1.1
faiss-cpu

